Install user account into container if asked for (with -u).
If you pass -u to lxc create, it will bind mount /home into the
container. Also, if you're running lxc with sudo, it will create a user
account in the container as that user, using the account details from
the host (e.g. you'll have the same password).

This makes sshing to the vserver quite nice as your normal user - if
you're in your own authorized_keys file of course.