Add 'autostart', 'stopall' and upstart config so containers can be started on boot.
If you pass -a to create, it will mark a container as wanting autostart
(or you can touch /var/lib/lxc/[name]/autostart).

The upstart jobs use the autostart and stopall commands.

Note that their implementations are serial, so starting many containers
will probably be slow - something to fix later.